requirements.txt
Cython==3.0.8
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
nltk==3.6.2
matplotlib==3.4.2
python-dotenv==0.19.2
scipy==1.7.1
xgboost==1.5.0